Add ScenePosition innerclass to help with movement
adds an innerclass to Stage which handles the movement. It provides
useful functions such as increment and decrement movement, while
handling all the modular arithmetic in the background.

It is used simply by: First incrementing or decrementing your position
until your happy with it, and then providing it to scene to be used to
discover your position.

These functions also return a boolean if they wrap around, which allows
for actors to check if that happens