ActorSprite now wraps, ActorPositioning uses Scene
ActorSprite now has a classname and correctly wraps around the edges
with the new functions provided based on which direction it is wrapping
around. Since offstage left exists, but offstage right doesn't the
direction you go determines your teleport order.

Additionally update the ActorPositioning to use the new
ScenePositioning, as well as refactoring the movement code in general to
be easier to work with in the new system.

This also meant changing the way keys were passed around. I opted for
separating movement and turning since I often wanted to check them
separately. I like the change.